backend/requirements.txt
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.1.2
anyio==4.0.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitarray==2.8.1
blinker==1.7.0
boltons @ file:///C:/ci_311/boltons_1677729932371/work
Brotli==1.1.0
brotlipy==0.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi @ file:///C:/ci_311/cffi_1676423759166/work
chainlink-feeds==0.2.9
charset-normalizer==3.3.2
click==8.1.7
colorama @ file:///C:/ci_311/colorama_1676422310965/work
comm==0.1.4
conda @ file:///C:/b/abs_b5q92oum8h/croot/conda_1689272261421/work
conda-content-trust @ file:///C:/ci_311/conda-content-trust_1676467587370/work
conda-libmamba-solver @ file:///C:/b/abs_5a9xvt2a3k/croot/conda-libmamba-solver_1685032355900/work/src
conda-package-handling @ file:///C:/b/abs_ce4_vcfd0y/croot/conda-package-handling_1685024800103/work
conda_package_streaming @ file:///C:/b/abs_88a7k_wmm1/croot/conda-package-streaming_1685019697115/work
configparser==6.0.0
cryptography @ file:///C:/b/abs_b28tzldafl/croot/cryptography_1686613103834/work
cytoolz==0.12.2
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==5.1.1
duckduckgo-search==3.9.3
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
executing==2.0.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hexbytes==0.3.1
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna @ file:///C:/ci_311/idna_1676424932545/work
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.5.0
jupyter_core==5.5.0
kaleido==0.2.1
libmambapy @ file:///C:/b/abs_f0tnqlkcbn/croot/mamba-split_1685993165748/work/libmambapy
lru-dict==1.2.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
menuinst @ file:///C:/ci_311/menuinst_1678730372782/work
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.25.2
oauthlib==3.2.2
packaging @ file:///C:/b/abs_ed_kb9w6g4/croot/packaging_1678965418855/work
pandas==2.0.3
parsimonious==0.9.0
parso==0.8.3
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
pluggy @ file:///C:/ci_311/pluggy_1676422178143/work
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.1
pycosat @ file:///C:/ci_311/pycosat_1676438455539/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.18.0
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL @ file:///C:/b/abs_de215ipd18/croot/pyopenssl_1678965319166/work
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil==2.8.2
python-dotenv==1.0.0
pytrends==4.9.2
pytz==2023.3
pyunormalize==15.0.0
pywin32==306
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rlp==3.0.0
rpds-py==0.9.2
ruamel.yaml @ file:///C:/ci_311/ruamel.yaml_1676439214109/work
sentinelhub==3.9.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlitedict==2.1.0
stack-data==0.6.3
streamlit==1.28.2
tenacity==8.2.3
tifffile==2023.9.26
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
toolz @ file:///C:/ci_311/toolz_1676431406517/work
tornado==6.3.3
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
traitlets==5.13.0
twilio==8.10.2
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
utm==0.7.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.9
web3==6.8.0
websockets==11.0.3
Werkzeug==3.0.1
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
yarl==1.9.2
zipp==3.17.0
zstandard==0.19.0
